Update testv3.html
# function SetLimit
This function is used to dynamically set the threshold of the minimum level sensor at which the system will detect it as nearly empt. Look at the arduino sketch for
more info.

# function SilosLevel
This function will dynamically display the current silos level on the meter in Figure 2.6.
Values are updated every 10 seconds in order to give arduino server enough time to process
the request/response. Look at the arduino sketch for more info